fix: address secret passphrase review findings
Keep overlong passphrase failures recoverable at startup and during optional legacy rewraps. Clarify credential-specific validation, avoid oversized per-frame width probes, and update the changelog and user stories.

+++ src/wallet_backend/single_key.rs
@@ -1851,7 +1851,7 @@ mod tests {
             .expect_err("over-long passphrase rejected");
         match err {
             TaskError::SingleKeyPassphraseTooLong { max } => {
-                assert_eq!(max, MAX_PASSPHRASE_LEN as u32);
+                assert_eq!(max, MAX_PASSPHRASE_LEN);
             }
             other => panic!("expected SingleKeyPassphraseTooLong, got {other:?}"),
         }
